+++ internal/adapter/repository/facility_respository.go
@@ -11,11 +11,11 @@ import (
 )
 
 type FacilityRepository interface {
-	CreateFacilityCoffeeShop(ctx context.Context, idFacility, idCoffeShop int) error
-	CodeForCreateFCS(ctx context.Context, code string) (int, error)
+	CreateFacilityCoffeeShop(ctx context.Context, idFacility, idCoffeShop int64) error
+	CodeForCreateFCS(ctx context.Context, code string) (int64, error)
 	CreateFacility(ctx context.Context, req entity.FacilityEntity) error
-	UpdateFacility(ctx context.Context, req entity.FacilityEntity, id int) error
-	DeleteFacility(ctx context.Context, id int) error
+	UpdateFacility(ctx context.Context, req entity.FacilityEntity, id int64) error
+	DeleteFacility(ctx context.Context, id int64) error
 	GetFacilities(ctx context.Context) ([]entity.FacilityEntity, error)
 }
 
@@ -24,7 +24,7 @@ type facilityRepository struct {
 }
 
 // DeleteFacility implements [FacilityRepository].
-func (f *facilityRepository) DeleteFacility(ctx context.Context, id int) error {
+func (f *facilityRepository) DeleteFacility(ctx context.Context, id int64) error {
 	var modelFacility model.Facility
 	err := f.db.Where("id = ?", id).Delete(&modelFacility).Error
 	if err != nil {
@@ -57,7 +57,7 @@ func (f *facilityRepository) GetFacilities(ctx context.Context) ([]entity.Facili
 }
 
 // UpdateFacility implements [FacilityRepository].
-func (f *facilityRepository) UpdateFacility(ctx context.Context, req entity.FacilityEntity, id int) error {
+func (f *facilityRepository) UpdateFacility(ctx context.Context, req entity.FacilityEntity, id int64) error {
 	modelFacility := model.Facility{
 		Name: req.Name,
 		Code: req.Code,
@@ -99,7 +99,7 @@ func (f *facilityRepository) CreateFacility(ctx context.Context, req entity.Faci
 }
 
 // CodeForCreateFCS implements [FacilityRepository].
-func (f *facilityRepository) CodeForCreateFCS(ctx context.Context, code string) (int, error) {
+func (f *facilityRepository) CodeForCreateFCS(ctx context.Context, code string) (int64, error) {
 	var modelFacility model.Facility
 	err := f.db.Where("code = ?", code).Find(&modelFacility).Error
 	if err != nil {
@@ -111,7 +111,7 @@ func (f *facilityRepository) CodeForCreateFCS(ctx context.Context, code string) 
 }
 
 // CreateFacilityCoffeeShopRelation implements [FacilityRepository].
-func (f *facilityRepository) CreateFacilityCoffeeShop(ctx context.Context, idFacility, idCoffeShop int) error {
+func (f *facilityRepository) CreateFacilityCoffeeShop(ctx context.Context, idFacility, idCoffeShop int64) error {
 	modelCSFacility := model.CoffeeShopFacility{
 		CoffeeShopID: idCoffeShop,
 		FacilityID:   idFacility,
