+++ internal/core/domain/model/coffeeshop_facility_model.go
@@ -1,10 +1,10 @@
 package model
 
 type CoffeeShopFacility struct {
-	ID           int `gorm:"id"`
-	CoffeeShopID int `gorm:"coffee_shop_id"`
+	ID           int64 `gorm:"id"`
+	CoffeeShopID int64 `gorm:"coffee_shop_id"`
 	CoffeeShop CoffeeShop `gorm:"foreignKey:CoffeeShopID"`
-	FacilityID   int `gorm:"facility_id"`
+	FacilityID   int64 `gorm:"facility_id"`
 	Facility Facility `gorm:"foreignKey:FacilityID"`
 }
