+++ internal/core/domain/model/coffeeshop_image_model.go
@@ -3,8 +3,8 @@ package model
 import "time"
 
 type CoffeeShopImage struct {
-	ID          int `gorm:"id"`
-	CoffeeShopID int `gorm:"coffee_shop_id"`
+	ID          int64 `gorm:"id"`
+	CoffeeShopID int64 `gorm:"coffee_shop_id"`
 	Image       string `gorm:"image"`
 	IsPrimary   bool `gorm:"is_primary"`
 	CreatedAt   time.Time `gorm:"created_at"`
