+++ internal/core/domain/model/coffeeshop_model.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type CoffeeShop struct {
-	ID         int `gorm:"id"`
+	ID         int64 `gorm:"id"`
 	Name       string `gorm:"name"`
 	Address    string `gorm:"address"`
 	Latitude   float64 `gorm:"latitude"`
@@ -11,14 +11,14 @@ type CoffeeShop struct {
 	OpenTime   string `gorm:"open_time"`
 	CloseTime string `gorm:"close_time"`
 	Instagram string `gorm:"instagram"`
-	CreatedByID int `gorm:"created_by_id"`
+	CreatedByID int64 `gorm:"created_by_id"`
 	UserCreate User `gorm:"foreignKey:CreatedByID"`
-	UpdatedByID int `gorm:"updated_by_id"`
+	UpdatedByID int64 `gorm:"updated_by_id"`
 	UserUpdate User `gorm:"foreignKey:UpdatedByID"`
 	CreatedAt time.Time `gorm:"created_at"`
 	UpdatedAt time.Time `gorm:"updated_at"`
 	IsActive bool `gorm:"is_active"`
-	CategoryID int `gorm:"category_id"`
+	CategoryID int64 `gorm:"category_id"`
 	Category Category `gorm:"foreignKey:CategoryID"`
 	Images []CoffeeShopImage `gorm:"foreignKey:CoffeeShopID"`
 	CoffeeShopFacility []CoffeeShopFacility `gorm:"foreignKey:CoffeeShopID"`
