+++ internal/core/domain/model/facility_model.go
@@ -1,7 +1,7 @@
 package model
 
 type Facility struct {
-	ID int `gorm:"id"`
+	ID int64 `gorm:"id"`
 	Code string `gorm:"code"`
 	Name string `gorm:"name"`
 }
